requirements.txt
# ============================================
# AI Phishing Detector SaaS — Dependencies
# ============================================

# --- Backend ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# --- ML & Data ---
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
scipy==1.13.1

# --- Text Processing ---
beautifulsoup4==4.12.3
tldextract==5.1.2

# --- Auth & Security ---
PyJWT==2.8.0
bcrypt==4.2.1
python-dotenv==1.0.1

# --- Frontend (Streamlit) ---
streamlit==1.36.0
requests==2.32.3

# --- Testing ---
pytest==8.2.2
httpx==0.27.0

# --- Utilities ---
pydantic[email]==2.7.4
pydantic-settings==2.3.4
